refactor: extract descriptors reader to a package
